+++ changelog.md
@@ -112,8 +112,8 @@
 - Added ``macros.copyLineInfo`` to copy lineInfo from other node.
 - Added ``system.ashr`` an arithmetic right shift for integers.
 
-- Added `split`, `splitWhitespace`, `size` procs and iterators
-  to `unicode.nim`.
+- Added `split`, `splitWhitespace`, `size`, `alignLeft`, `align`,
+  `strip`, `repeat` procs and iterators to `unicode.nim`.
 
 ### Library changes